--- src/helpers/category.ts
@@ -1,11 +1,0 @@
-import { ICategory } from "@/interfaces/Types";
-
-const categoryToPreLoad: ICategory[] = [
-  { id: "All", name: "All", icon: "LoaderPinwheel" },
-  { id: "Burgers", name: "Burgers", icon: "Sandwich" },
-  { id: "Drinks", name: "Drinks", icon: "CupSoda" },
-  { id: "Desserts", name: "Desserts", icon: "IceCreamCone" },
-  { id: "Sides", name: "Sides", icon: "Salad" },
-];
-
-export default categoryToPreLoad;
